extractors/txt/requirements.txt
pika==1.3.1
nltk==3.8.1
rake-nltk==1.0.6
python-dotenv==1.0.0 